+++ src/hlsl++_unit_tests_vector_int.cpp
@@ -586,10 +586,10 @@ void RunUnitTestsVectorInt()
 
 	hlslpp_unit::BeginTest("Integer Float Conversions");
 	{
-		float f1 = (rand() % 1000) / 100.0f; float f5 = (rand() % 1000) / 100.0f; float f9 = (rand() % 1000) / 100.0f; //float f13 = (rand() % 1000) / 100.0f; float f17 = (rand() % 1000) / 100.0f;
-		float f2 = (rand() % 1000) / 100.0f; float f6 = (rand() % 1000) / 100.0f; float f10 = (rand() % 1000) / 100.0f; //float f14 = (rand() % 1000) / 100.0f; float f18 = (rand() % 1000) / 100.0f;
-		float f3 = (rand() % 1000) / 100.0f; float f7 = (rand() % 1000) / 100.0f; //float f11 = (rand() % 1000) / 100.0f; float f15 = (rand() % 1000) / 100.0f; float f19 = (rand() % 1000) / 100.0f;
-		float f4 = (rand() % 1000) / 100.0f; float f8 = (rand() % 1000) / 100.0f; //float f12 = (rand() % 1000) / 100.0f; float f16 = (rand() % 1000) / 100.0f; float f20 = (rand() % 1000) / 100.0f;
+		float f1 = (float)(rand() % 1000) / 100.0f; float f5 = (float)(rand() % 1000) / 100.0f; float f9 = (float)(rand() % 1000) / 100.0f; //float f13 = (rand() % 1000) / 100.0f; float f17 = (rand() % 1000) / 100.0f;
+		float f2 = (float)(rand() % 1000) / 100.0f; float f6 = (float)(rand() % 1000) / 100.0f; float f10 = (float)(rand() % 1000) / 100.0f; //float f14 = (rand() % 1000) / 100.0f; float f18 = (rand() % 1000) / 100.0f;
+		float f3 = (float)(rand() % 1000) / 100.0f; float f7 = (float)(rand() % 1000) / 100.0f; //float f11 = (rand() % 1000) / 100.0f; float f15 = (rand() % 1000) / 100.0f; float f19 = (rand() % 1000) / 100.0f;
+		float f4 = (float)(rand() % 1000) / 100.0f; float f8 = (float)(rand() % 1000) / 100.0f; //float f12 = (rand() % 1000) / 100.0f; float f16 = (rand() % 1000) / 100.0f; float f20 = (rand() % 1000) / 100.0f;
 
 		// Initialization
